+++ src/main/java/com/arangodb/ArangoEdgeCollectionAsync.java
@@ -36,7 +36,7 @@ import com.arangodb.model.EdgeUpdateOptions;
  * @see <a href="https://docs.arangodb.com/current/HTTP/Gharial/Edges.html">API Documentation</a>
  * @author Mark Vollmary
  */
-public interface ArangoEdgeCollectionAsync {
+public interface ArangoEdgeCollectionAsync extends ArangoSerializationAccessor {
 
 	/**
 	 * The the handler of the named graph the edge collection is within