+++ src/main/java/com/arangodb/ArangoVertexCollectionAsync.java
@@ -36,7 +36,7 @@ import com.arangodb.model.VertexUpdateOptions;
  * @see <a href="https://docs.arangodb.com/current/HTTP/Gharial/Vertices.html">API Documentation</a>
  * @author Mark Vollmary
  */
-public interface ArangoVertexCollectionAsync {
+public interface ArangoVertexCollectionAsync extends ArangoSerializationAccessor {
 
 	/**
 	 * The the handler of the named graph the edge collection is within